Fix bug in range of supported timezones in `SlackUser` (#61)
This changes the type of `tz_offset` from `i16` to `i32`.

`tz_offset` is the number of seconds to offset UTC time by.
If the offset is greater than 9 hours, it is outside i16 range when
represented in seconds.

+++ src/models/src/common/user.rs
@@ -16,7 +16,7 @@ pub struct SlackUser {
     pub flags: SlackUserFlags,
     pub tz: Option<String>,
     pub tz_label: Option<String>,
-    pub tz_offset: Option<i16>,
+    pub tz_offset: Option<i32>,
     pub updated: Option<SlackDateTime>,
     pub deleted: Option<bool>,
     pub color: Option<SlackColor>,
